requirements.txt
presidio_analyzer >= 2.2.32

pii-data >= 0.4.0, <1.0.0
pii-extract-base >= 0.4.0, <1.0.0
